requirements.txt
streamlit>=1.31
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
matplotlib>=3.7
pyarrow>=14.0
